update powertimer.py: new method update_loop() and vars self.loop, self.loopEvent and self.loopAlways
basically what I tried is to set a loopEvent when the loop is finished. I created also the self.loopAlways depending on the value of loop so the update_loop() method can track the status of the loop. It's been a tiring day so I'm just commiting a few changes before I screw up everything.